added rosbag inspector widget. unfinished
