+++ src/ui/window.cpp
@@ -58,6 +58,7 @@ namespace vk {
         // TODO: Move to ContentRegistry and call get_entries() to get the list of views to render in a loop
         this->view_corner_detector.draw_content();
         this->view_recorder.draw_content();
+        this->view_rosbag_inspector.draw_content();
     }
 
     void Window::frame_end() {
